# Add optimization for quote_char: nil