--- a/lib/csv/parser.rb
+++ b/lib/csv/parser.rb
@@ -227,6 +227,33 @@ class CSV
       last_line
     end
 
+    def parse_quote_character_nil(&block)
+      while true
+        return nil unless value = @input.gets(@row_separator)
+        next if @skip_lines and skip_line?(value)
+
+        value.chomp!
+
+        row = []
+        if value.empty?
+          next if @skip_blanks
+        else
+          columns = if @column_separator == " "
+            value.split(@column_end, -1)
+          else
+            value.split(@column_separator, -1)
+          end
+          columns.each do |column|
+            row << (column.empty? ? nil : column)
+          end
+        end
+        @last_line = value
+
+        emit_row(row, &block)
+      end
+      @parsed = true
+    end
+
     def parse(&block)
       return to_enum(__method__) unless block_given?
 
@@ -242,6 +269,8 @@ class CSV
 
       row = []
       begin
+        return parse_quote_character_nil(&block) if quote_character.nil?
+
         @scanner = build_scanner
         skip_needless_lines
         start_row
@@ -304,7 +333,7 @@ class CSV
       prepare_regexp
       prepare_line
       prepare_header
-      prepare_parser
+      prepare_parser if @options[:quote_character]
     end
 
     def prepare_variable
@@ -335,20 +364,26 @@ class CSV
       @column_separator = @options[:column_separator].to_s.encode(@encoding)
       @row_separator =
         resolve_row_separator(@options[:row_separator]).encode(@encoding)
-      @quote_character = @options[:quote_character].to_s.encode(@encoding)
-      if @quote_character.length != 1
-        raise ArgumentError, ":quote_char has to be a single character String"
-      end
+
+
       @backslash_character = "\\".encode(@encoding)
+      if @options[:quote_character].nil?
+        @quote_character = nil
+      else
+        @quote_character = @options[:quote_character].to_s.encode(@encoding)
+        if @quote_character.length != 1
+          raise ArgumentError, ":quote_char has to be a single character String"
+        end
+        escaped_quote_character = Regexp.escape(@quote_character)
+        @escaped_quote = Regexp.new(escaped_quote_character)
+        @backslash_quote_character = @backslash_character + escaped_quote_character
+      end
 
       escaped_column_separator = Regexp.escape(@column_separator)
       escaped_first_column_separator = Regexp.escape(@column_separator[0])
       escaped_row_separator = Regexp.escape(@row_separator)
-      escaped_quote_character = Regexp.escape(@quote_character)
       escaped_backslash_character = Regexp.escape(@backslash_character)
       @escaped_backslash = Regexp.new(escaped_backslash_character)
-      @escaped_quote = Regexp.new(escaped_quote_character)
-      @backslash_quote_character = @backslash_character + escaped_quote_character
 
       skip_lines = @options[:skip_lines]
       case skip_lines
@@ -385,23 +420,26 @@ class CSV
         @row_ends = nil
       end
 
-      @quotes = Regexp.new(escaped_quote_character +
-                           "+".encode(@encoding))
-      if @backslash_quote
-        @quoted_value = Regexp.new("[^".encode(@encoding) +
-                                   escaped_quote_character +
-                                   escaped_backslash_character +
-                                   "]+".encode(@encoding))
-      else
-        @quoted_value = Regexp.new("[^".encode(@encoding) +
-                                   escaped_quote_character +
-                                   "]+".encode(@encoding))
+      if @options[:quote_character]
+        @quotes = Regexp.new(escaped_quote_character +
+                             "+".encode(@encoding))
+        if @backslash_quote
+          @quoted_value = Regexp.new("[^".encode(@encoding) +
+                                     escaped_quote_character +
+                                     escaped_backslash_character +
+                                     "]+".encode(@encoding))
+        else
+          @quoted_value = Regexp.new("[^".encode(@encoding) +
+                                     escaped_quote_character +
+                                     "]+".encode(@encoding))
+        end
       end
+
       if @liberal_parsing
         @unquoted_value = Regexp.new("[^".encode(@encoding) +
                                      escaped_first_column_separator +
                                      "\r\n]+".encode(@encoding))
-      else
+      elsif @options[:quote_character]
         @unquoted_value = Regexp.new("[^".encode(@encoding) +
                                      escaped_quote_character +
                                      escaped_first_column_separator +