I implement https://github.com/ruby/csv/issues/56

This implementation uses `line.split(column_separator)` to parse. As a result, it got 2.7x faster for `quote_char: nil` case.
Also, as the number of columns to parse increases, `quote_char: nil` is more advantageous in terms of performance.

```
benchmark:
$ rake benchmark:parse_quote_char_nil

(Case of 50 columns)
Comparison:
      quote_char_nil:        76.4 i/s 
         encode_sjis:        52.9 i/s - 1.45x  slower
        encode_utf-8:        51.5 i/s - 1.48x  slower
       col_sep_space:        48.3 i/s - 1.58x  slower
  without_quote_char:        28.3 i/s - 2.70x  slower

(Case of 10 columns)
Comparison:
      quote_char_nil:       265.4 i/s 
         encode_sjis:       206.5 i/s - 1.29x  slower
        encode_utf-8:       196.9 i/s - 1.35x  slower
       col_sep_space:       172.6 i/s - 1.54x  slower
  without_quote_char:       120.9 i/s - 2.20x  slower
```

If col_sep is a single byte blank character `" "`, `String#split` is divided by a blank string after excluding leading and trailing blanks.
In order to support the following cases, use regular expression in `String#split` only when col_sep is a space character.

```ruby
CSV.parse("a b  d", col_sep: " ", quote_char: nil)
#=> [["a", "b", nil, "d"]]
```

However, when using regular expressions, it is about 1.5 times slower than using strings. The above col_sep_space is the benchmark result. Even in this case it is faster than the existing parse implementation.

Unfortunately, This implementation currently does not support stream.