@284km 2019-03-12
benchmark result is below. 3.77x faster!
https://travis-ci.org/ruby/csv/jobs/505165753

```
(Case of 50 columns)
      quote_char_nil:       103.7 i/s 
       col_sep_space:        56.2 i/s - 1.84x  slower
  without_quote_char:        27.5 i/s - 3.77x  slower

(Case of 10 columns)
      quote_char_nil:       309.5 i/s 
       col_sep_space:       168.6 i/s - 1.84x  slower
  without_quote_char:       122.4 i/s - 2.53x  slower
```